Media sources and support for multiple viewers
Refactors PeerConnection to use audio and video tracks provided
by a media source. File and V4L2 media sources are implemented.
Audio support is not implemented.

As V4L2 device may only be opened by a single reader, need to
copy frames to multiple readers in the application. Created a
MediaSource interface, which specifies a GetTrack() function for
returning a track to pass to PeerConnection.

Also, fixed improper use of select() in v4l2.Read() when device is
closed -- this became apparent with refactoring into a video track,
and adding proper end-of-track error propagation. The select()
call would block on the file descriptor, and while it was blocked,
close() would be called on the file description when the peer closed
the connection. POSIX labels this is labeled as 'undefined behavior'.
Linux was returning a 'bad file descriptor error', but if the same
file descriptor happened to be used for something else, select()
would block indefinitely. Replaced select() entirely, making the
v4l2 DQBUF ioctl() block -- when the device is closed while blocked,
the ioctl returns a -EINVAL error, which is then propagated up as
an io.EOF.

Discovered that STAP-A timestamp must match first frame after
STAP-A frame, else video will not start.

+++ examples/v4l2/demo.go
@@ -46,7 +46,7 @@ func main() {
 	if bitrate, err := strconv.Atoi(flagBitrate); err != nil {
 		panic(err)
 	} else {
-		dev.SetBitrate(bitrate)
+		dev.SetBitrate(uint(bitrate))
 	}
 
 	// Start
